backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
supabase>=2.0.0
pydantic>=2.6.0
python-dotenv>=1.0.1

pytest>=7.4.0
httpx>=0.26.0
pytest-asyncio>=0.23.0
resend>=0.1.0
psycopg2-binary>=2.9.9

pytz==2024.1
